packages/paperlint/pyproject.toml
[project]
name = "paperlint"
version = "0.2.0"
description = "LLM pipeline for WG21 paper defect discovery"
requires-python = ">=3.12"
license = {text = "BSL-1.0"}
authors = [{name = "Sergio DuBois", email = "sentientsergio@gmail.com"}]
dependencies = [
    "paperstore",
    "mailing",
    "tomd",
    "openai",
    "python-dotenv",
    "requests",
]

[project.scripts]
paperlint = "paperlint.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paperlint"]
